Moved api settings in config.py
